--- LawOf100.Web/Features/DayState.cs
@@ -1,28 +1,0 @@
-﻿using LawOf100.Features.Features.Entities;
-using Sparc.Core;
-using Sparc.Features;
-using System.Linq;
-using System.Linq.Dynamic.Core;
-
-namespace LawOf100.Features.Features
-{
-    public class DayState : PublicFeature<string, List<Progression>>
-    {
-        
-        public DayState(IRepository<Progression> daystates)
-        {
-            DayStates = daystates;
-        }
-
-        public IRepository<Progression> DayStates { get; }
-
-        public override async Task<List<Progression>> ExecuteAsync(string currentstate)
-        {
-            var daystates = await DayStates.Query
-                .Where(x => x.DayState == currentstate)
-                .ToListAsync();
-
-            return daystates;
-        }
-    }
-}
