+++ LawOf100.Web/Habits/Entities/Recurrence.cs
@@ -1,0 +1,5 @@
+﻿namespace LawOf100.Features.Habits.Entities;
+
+public class Recurrence
+{
+}
